Move app and user reducer location
