+++ src/renderer/redux/middleware/imagingDeviceMiddleware.ts
@@ -4,8 +4,8 @@ import {
   deviceIdentifierToDevice,
   ImagingDevice,
 } from "../../services/imagingDevice/ImagingDevice";
-import { setCurrentDevice, setDeviceStreaming } from "../reducers/app/reducer";
-import { setCurrentDeviceFromId } from "../reducers/app/thunks";
+import { setCurrentDevice, setDeviceStreaming } from "../app/reducer";
+import { setCurrentDeviceFromId } from "../app/thunks";
 import { RootState } from "../store";
 import { withLoader } from "../utils";
