+++ src/renderer/redux/store.ts
@@ -1,8 +1,8 @@
 import { applyMiddleware, combineReducers, createStore } from "redux";
 import thunk from "redux-thunk";
+import appReducer from "./app/reducer";
 import { createCaptureMiddleware } from "./middleware/imagingDeviceMiddleware";
-import appReducer from "./reducers/app/reducer";
-import userPreferencesReducer from "./reducers/userPreferences/reducer";
+import userPreferencesReducer from "./userPreferences/reducer";
 export type RootState = ReturnType<typeof rootReducer>;
 
 const rootReducer = combineReducers({
