+++ src/renderer/redux/userPreferences/reducer.ts
